move RunResult out of SSH transport module
