+++ bundlewrap/transports/__init__.py
@@ -1,0 +1,17 @@
+from ..utils import cached_property
+from ..utils.text import force_text
+
+
+class RunResult(object):
+    def __init__(self):
+        self.return_code = None
+        self.stderr = None
+        self.stdout = None
+
+    @cached_property
+    def stderr_text(self):
+        return force_text(self.stderr)
+
+    @cached_property
+    def stdout_text(self):
+        return force_text(self.stdout)
