+++ bundlewrap/transports/ssh.py
@@ -8,8 +8,8 @@ from subprocess import Popen, PIPE
 from threading import Event, Thread
 from os import close, environ, pipe, read, setpgrp
 
+from . import RunResult
 from ..exceptions import RemoteException
-from ..utils import cached_property
 from ..utils.text import force_text, LineBuffer, mark_for_translation as _, randstr
 from ..utils.ui import io
 
@@ -69,21 +69,6 @@ def download(
             host=hostname,
             path=remote_path,
         ))
-
-
-class RunResult(object):
-    def __init__(self):
-        self.return_code = None
-        self.stderr = None
-        self.stdout = None
-
-    @cached_property
-    def stderr_text(self):
-        return force_text(self.stderr)
-
-    @cached_property
-    def stdout_text(self):
-        return force_text(self.stdout)
 
 
 def run(
